Localised date format
* Localised date format based on the user's browser language.
* Added jsdom as the jest environment.
* Moved date related functions from main.js into date-helper.js.
* Added tests for formatDate to make sure the date formatting is working correctly based on the user's browser language.
