+++ jest.config.json
@@ -1,0 +1,3 @@
+{
+  "testEnvironment": "jsdom"
+}
