+++ package.json
@@ -14,6 +14,7 @@
     "@babel/preset-env": "^7.19.3",
     "babel-jest": "^29.1.2",
     "jest": "^29.1.2",
+    "jest-environment-jsdom": "^29.1.2",
     "prettier": "2.7.1",
     "vite": "^3.0.7"
   },
